✨ feat: ProjectPage / ProjectDetailPage 분리
