+++ app/src/routes/ROUTES.ts
@@ -2,8 +2,8 @@ export const ROUTES = {
   ROOT: '/',
   FTOAUTH: '/ftoauth',
   PROFILE: '/profile/:username',
-  PROJECT_ROOT: '/project',
-  PROJECT: '/project/:projectName',
+  PROJECT: '/project',
+  PROJECT_DETAIL: '/project/:projectName',
   LEADERBOARD: '/leaderboard',
   EVALLOG: '/evallog',
   LOGIN: '/login',
